\name{udhr.demo}
\docType{data}
\alias{udhr.demo}
\title{Universal Declaration of Human Rights demo corpus}
\description{
  This data set provides the English and Italian processed translations of the
  Universal Declaratation of Human Rights. The pre-processing involved
  the removal of punctuation, newlines, leading copyright lines, article
  numbers, conversion to lowercase and trimming of multiple spaces to one.
}
\usage{udhr.demo}
\format{A \code{list} containing two \code{character vectors}.}
\source{UDHR in Unicode \url{http://unicode.org/udhr/}}
\keyword{datasets}